Move initialization to Index.jsx
I need to be able to render the App component on load, and I was
previously trying to do that in the Initializer class. However, that
required importing the App component there. It seemed redundant, so I
opted to just do that in the Index file. After this, I should not need
to add anything to the Index file, as the App component will import and
use the other (future) components as needed.